Adding trait implementation
- requires instance: WIP
- catches default behaviour: Unsupported

Simplifying SftpServer trait async and the Handle are out for now

I would like to keep those but since I want a structured of SftpServer to keep it's own internal state, I could not use async or a Handle with a trait definition. It is unfortunately unstable.

Still a WIP but now a client can PUT a small file and its content will be displayed in debug.

For now the buffers overflow on long SSH_FXP_WRITE packets and fail after trying to decode in the next loop iteration more of the same packet content.

The previous sftpserver trait has been stored in asyncsftpserver.rs

Simple random data files generation for big write testing

Adding SFTP Packet Header and SSH_FXP_WRITE offset definitions

Simplified Main Demo SFTP loop and reduced the SFTP buffers size

WIP: Working on long WRITE packets processing

WIP: Processing long Write Requests

For now the flow of receiving SFTP Write request longer than the buffer is going in the right direction. I have been able to process up to 64kB of data. Still this is not complete since I need to:

- Write the data received and check that the data sent and received has not been altered
- Issue: Writes with files over 64kB fail
- Reshape the file handle exchanged from the demoserver so it has a fixed length and it is compact
- Refactor sftphandle

Adding basic obscuring of handler

This forces the SftpServer to use obscured handles that are size constrain.

It also introduces the handle manager, that is a helper that the SFTP server implementer can use for dealing with the obscured handle and a private handle, that can contain all the information that the local server requires.

Minimal SFTP Server implementation of write 64kB write pass the test

File sent and written are identical. That is a good start.

Next I will find what is wrong passed 64kB

Troubleshooting the 65kB limit: There are two write requests

I was assuming that the client would wait for a `SSH_FXP_STATUS` Ok, but it does not. I need to handle a second request segment in the same SftpSource