Handing consecutive request in a buffer with some problems
The issue that is left to fix is what happen when the second request is trunked and cannot be decoded.

I am contemplating creating a "truncated" request buffer that would be used when the request cannot be decoded because requires extra data to be received in the next process call.

Apart from this, the current code is able to process fragmented and consecutive SSH_FXP_WRITE for files with occasional problems caused by the issue detailed.