[skip ci] WIP: BUG The EOF needs to be sent once the readdir has sent all the items and is asked again. Anything else fails
This is point I have not considered. I could have sent one item at a time, with different challenges.

In my opinion I should add this behavior to the SftpServer implementation, but that passes the responsibility to the library user and the idea is making the user not needing to care too much about the protocol implementation

[skip ci] WIP: Typo and cleanup commented out code