[skip ci] After finally applying the suggested fix responses stall in 100MB get
Will provide further logs in a comment.
Improved log scripts.

In summary, the communication stalls because the ChanOut.write at some point does not return after being called.

The reception if packets does not show any anomaly.

[skip ci] Recording better logs

Coordinating pcap files, test script output, demo program output and strace (only network traffic)

[skip ci] Exploring logs: Found important issue in OutputChannelHandler

Not all the data that the producer tries to send reach the consumer, that is the one that writes the data in the ChannelOut.

This is likely to be the responsible for the stuck communication.

Apart from that, I added more testing and fixed issue in Proto::Data (SSH_FXP_DATA). It had too many fields.

More details to come

[skip ci] Fixing test for Data roundtrip and updating Cargo.lock

[skip ci] Removed env-logger dependency for sftp

It was introduced for read_packet_from_file.rs and it is no longer needed

[skip ci] Extracting sftpoutputproducer bytes for analysis

At this point I can see that on stall, the backpressure from the ChanOut blocks the data put in the pipe.

I took measures in a previous commit (5d1c1f9424aa4ec5eb00450e4c21e2dde96eaaf9) to guarantee that for SSH_FXP_DATA and SSH_FXP_NAME the announced length and real length matches.
- SftpServer Unit tests to check the length in header and data sent
- Runtime checks for ReadDir, Read and PathInfo requests

I am going to start digging in the ChanOut Write process

[skip ci] Replacing write_all with write loop with no change

This gives us visibility on when the ChanOut won't take all the bytes. It hapens regularly and that is expected. Does not provide extra visibility on the problem.

I will leave it like this for now

[skip ci] log_demo_sftp_with_tests: Adding automatic TX/RX data extraction

testing: Adding a script to merge client and server logs based on timestamp