Config num_enum for no_std
Fixing format and orphan use

Fixing docs format and docs for macros

Adding sftp to the CI fixing and passing it

Changing all SftpServer signatures to async (or impl Future with async move block)

I have done this in the view that all that operations might need to run async code.

Demo changed accordingly and test passing

demo/sftp/std readme, comments and added verbosity

Provides an overview of what you find in this demo and instructions on how to use it.

It also includes some warnings about security risks.

The SftpServer trait operations offer some info logs to illustrate the flow of the operations.

new feature: long-paths-1024, long-paths-4096

By default sunset-sftp will only allow paths up to 256 character long but this can be adjusted with the mention flags.

The motivation for this is keeping a small RequestHandler in the SftpHandler

applying the MAX_REQUEST_LEN and improving the signature of SftpHandler

Cleaning and completing TODOs, removing old files